refactors shadowcontainer and homepage component
