+++ src/main/components/UI/organisms/ShadowContainer/index.tsx
@@ -1,31 +1,21 @@
 import * as React from "react";
-import { DropFile } from "../DropFile";
 import styles from "./index.module.css";
 
 interface ShadowContainerDef {
     childContent: JSX.Element;
-}
-
-interface displayCompsDef {
-    [key: number]: () => JSX.Element;
+    current: number;
 }
 
 export const ShadowContainer = (props: ShadowContainerDef) => {
-    const [current, setCurrent] = React.useState(1);
-
-    const displayComps: displayCompsDef = {
-        1: () => <DropFile changeScreen={setCurrent} />,
-    };
-
     return (
         <article
             className={
-                current === 1
+                props.current === 1
                     ? `${styles.shdContScreenOne} ${styles.shdCont}`
                     : styles.shdCont
             }
         >
-            {displayComps[current]()}
+            {props.childContent}
         </article>
     );
 };
